+++ sdk-extensions/autoconfigure-spi/src/main/java/io/opentelemetry/sdk/autoconfigure/spi/traces/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * Java SPI (Service Provider Interface) for implementing extensions to SDK autoconfiguration of
+ * traces.
+ */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.autoconfigure.spi.traces;
+
+import javax.annotation.ParametersAreNonnullByDefault;
