+++ sdk-extensions/autoconfigure/src/main/java/io/opentelemetry/sdk/autoconfigure/spi/metrics/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * Java SPI (Service Provider Interface) for implementing extensions to SDK autoconfiguration of
+ * metrics.
+ */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.autoconfigure.spi.metrics;
+
+import javax.annotation.ParametersAreNonnullByDefault;
